+++ __tests__/unit/templates.test.js
@@ -15,7 +15,6 @@ const {
     _getExposedEntityTypes,
     _propagateORDVisibility,
     _handleVisibility,
-    isPrimaryDataProductService,
 } = require("../../lib/templates");
 
 const Logger = require("../../lib/logger");
@@ -138,62 +137,6 @@ describe("visibility handling", () => {
             groupId: "sap.cds:service:customer.testNamespace:MyService",
             groupTypeId: "sap.cds:service",
             title: "My Service",
-        });
-    });
-
-    describe("isPrimaryDataProductService", () => {
-        it("returns true for @DataIntegration.dataProduct.type: 'primary'", () => {
-            const serviceDefinition = { "@DataIntegration.dataProduct.type": "primary" };
-            expect(isPrimaryDataProductService(serviceDefinition)).toBe(true);
-        });
-
-        it("returns false for @DataIntegration.dataProduct.type: 'secondary'", () => {
-            const serviceDefinition = { "@DataIntegration.dataProduct.type": "secondary" };
-            expect(isPrimaryDataProductService(serviceDefinition)).toBe(false);
-        });
-
-        it("returns true for @data.product with truthy value", () => {
-            const serviceDefinition = { "@data.product": true };
-            expect(isPrimaryDataProductService(serviceDefinition)).toBe(true);
-        });
-
-        it("returns true for @data.product with any truthy value", () => {
-            const serviceDefinition = { "@data.product": "yes" };
-            expect(isPrimaryDataProductService(serviceDefinition)).toBe(true);
-        });
-
-        it("returns false for @data.product with falsy value", () => {
-            const serviceDefinition = { "@data.product": false };
-            expect(isPrimaryDataProductService(serviceDefinition)).toBe(false);
-        });
-
-        it("returns false for service with no data product annotations", () => {
-            const serviceDefinition = { "@title": "Regular Service" };
-            expect(isPrimaryDataProductService(serviceDefinition)).toBe(false);
-        });
-
-        it("returns true when both annotations are present - @DataIntegration.dataProduct.type takes precedence", () => {
-            const serviceDefinition = {
-                "@DataIntegration.dataProduct.type": "primary",
-                "@data.product": false,
-            };
-            expect(isPrimaryDataProductService(serviceDefinition)).toBe(true);
-        });
-
-        it("returns true when both annotations are present with @data.product truthy", () => {
-            const serviceDefinition = {
-                "@DataIntegration.dataProduct.type": "secondary",
-                "@data.product": true,
-            };
-            expect(isPrimaryDataProductService(serviceDefinition)).toBe(true);
-        });
-
-        it("returns false when both annotations are present with falsy values", () => {
-            const serviceDefinition = {
-                "@DataIntegration.dataProduct.type": "secondary",
-                "@data.product": false,
-            };
-            expect(isPrimaryDataProductService(serviceDefinition)).toBe(false);
         });
     });
 });
@@ -628,11 +571,10 @@ describe("templates", () => {
                 @EndUserText.label: 'This is test MyService event title'
                 service MyService { }
             `);
-            const eventResourceTemplate = createEventResourceTemplate(
-                linkedModel.definitions[serviceName],
-                appConfig,
-                ["sap.test.cdsrc.sample:package:test-event:v1", "sap.test.cdsrc.sample:package:test-api:v1"],
-            );
+            const eventResourceTemplate = createEventResourceTemplate(linkedModel.definitions[serviceName], appConfig, [
+                "sap.test.cdsrc.sample:package:test-event:v1",
+                "sap.test.cdsrc.sample:package:test-api:v1",
+            ]);
 
             expect(eventResourceTemplate).toBeInstanceOf(Array);
             expect(eventResourceTemplate[0].title).toEqual("This is test MyService event title");
@@ -661,11 +603,7 @@ describe("templates", () => {
                 "sap.test.cdsrc.sample:package:test-event:v1",
                 "sap.test.cdsrc.sample:package:test-api:v1",
             ];
-            const eventResourceTemplate = createEventResourceTemplate(
-                srvDefinition,
-                appConfig,
-                packageIds,
-            );
+            const eventResourceTemplate = createEventResourceTemplate(srvDefinition, appConfig, packageIds);
 
             expect(eventResourceTemplate).toBeInstanceOf(Array);
             expect(eventResourceTemplate).toMatchSnapshot();
@@ -694,11 +632,7 @@ describe("templates", () => {
                 "sap.test.cdsrc.sample:package:test-event-internal:v1",
                 "sap.test.cdsrc.sample:package:test-api:v1",
             ];
-            const eventResourceTemplate = createEventResourceTemplate(
-                srvDefinition,
-                appConfig,
-                packageIds,
-            );
+            const eventResourceTemplate = createEventResourceTemplate(srvDefinition, appConfig, packageIds);
 
             expect(eventResourceTemplate).toBeInstanceOf(Array);
             expect(eventResourceTemplate).toMatchSnapshot();
@@ -729,11 +663,7 @@ describe("templates", () => {
                 "sap.test.cdsrc.sample:package:test-event:v1",
                 "sap.test.cdsrc.sample:package:test-api:v1",
             ];
-            const eventResourceTemplate = createEventResourceTemplate(
-                srvDefinition,
-                appConfig,
-                packageIds,
-            );
+            const eventResourceTemplate = createEventResourceTemplate(srvDefinition, appConfig, packageIds);
 
             expect(eventResourceTemplate).toBeInstanceOf(Array);
             expect(eventResourceTemplate).toMatchSnapshot();
