+++ lib/protocol-resolver.js
@@ -1,6 +1,7 @@
 const cds = require("@sap/cds");
 
 const Logger = require("./logger");
+const { isPrimaryDataProductService } = require("./common/utils");
 const {
     CAP_TO_ORD_PROTOCOL_MAP,
     ORD_ONLY_PROTOCOLS,
@@ -18,15 +19,11 @@ const {
  * - Rule C: Never produce [null] in entryPoints
  *
  * @param {Object} srvDefinition The service definition object.
- * @param {Object} options Configuration options.
- * @param {Function} options.isPrimaryDataProduct Strategy function to check if service is primary data product.
  * @returns {Array} Array with single {apiProtocol, entryPoints, hasResourceDefinitions} object, or empty array.
  */
-function resolveApiResourceProtocol(srvDefinition, options = {}) {
-    const { isPrimaryDataProduct = () => false } = options;
-
+function resolveApiResourceProtocol(srvDefinition) {
     // 1. Primary Data Product - early return
-    if (isPrimaryDataProduct(srvDefinition)) {
+    if (isPrimaryDataProductService(srvDefinition)) {
         return [
             {
                 apiProtocol: ORD_API_PROTOCOL.SAP_DATA_SUBSCRIPTION,
