+++ lib/templates.js
@@ -1,9 +1,11 @@
-const defaults = require("./defaults");
 const _ = require("lodash");
+const defaults = require("./defaults");
+
+const Logger = require("./logger");
+const { ensureAccessStrategies } = require("./access-strategies");
+const { isPrimaryDataProductService } = require("./common/utils");
+const { resolveApiResourceProtocol } = require("./protocol-resolver");
 const {
-    DATA_PRODUCT_ANNOTATION,
-    DATA_PRODUCT_SIMPLE_ANNOTATION,
-    DATA_PRODUCT_TYPE,
     DESCRIPTION_PREFIX,
     ENTITY_RELATIONSHIP_ANNOTATION,
     LEVEL,
@@ -20,9 +22,6 @@ const {
     ORD_API_PROTOCOL,
     MCP_RESOURCE_DEFINITION_TYPE,
 } = require("./constants");
-const Logger = require("./logger");
-const { ensureAccessStrategies } = require("./access-strategies");
-const { resolveApiResourceProtocol } = require("./protocol-resolver");
 
 function unflatten(flattedObject) {
     let result = {};
@@ -283,9 +282,7 @@ const createAPIResourceTemplate = (srvDefinition, appConfig, packageIds, accessS
     const ordExtensions = readORDExtensions(srvDefinition);
     const visibility = _handleVisibility(ordExtensions, srvDefinition, appConfig.env?.defaultVisibility);
     const packageId = _getPackageID(appConfig.ordNamespace, packageIds, ORD_RESOURCE_TYPE.api, visibility);
-    const protocolResults = resolveApiResourceProtocol(srvDefinition, {
-        isPrimaryDataProduct: isPrimaryDataProductService,
-    });
+    const protocolResults = resolveApiResourceProtocol(srvDefinition);
     const apiResources = [];
 
     // If no protocols were generated, skip this service
@@ -494,13 +491,6 @@ const createEventResourceTemplate = (srvDefinition, appConfig, packageIds, acces
         : [];
 };
 
-function isPrimaryDataProductService(serviceDefinition) {
-    return (
-        serviceDefinition[DATA_PRODUCT_ANNOTATION] === DATA_PRODUCT_TYPE.primary ||
-        !!serviceDefinition[DATA_PRODUCT_SIMPLE_ANNOTATION]
-    );
-}
-
 function _getExposedEntityTypes(definitionObj) {
     if (!definitionObj.entities) {
         return;
@@ -621,5 +611,4 @@ module.exports = {
     _getExposedEntityTypes,
     _propagateORDVisibility,
     _handleVisibility,
-    isPrimaryDataProductService,
 };
